backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=23.0.0",
    "aiohttp>=3.9.0",
    "alembic>=1.13.0",
    "asyncpg>=0.29.0",
    "beautifulsoup4>=4.12.0",
    "fastapi>=0.118.3",
    "openai>=2.3.0",
    "pydantic[email]>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pypdf2>=3.0.1",
    "python-docx>=1.1.0",
    "python-dotenv>=1.0.0",
    "qdrant-client>=1.15.1",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.37.0",
]
